Optimize image tests * skip test conditional on presence of PreAuthenticatedRequest url specifying image location * run image tests in parallel * consolidate custom data source test into scenario test
